requirements.txt
flask
flask-cors
torch
torchvision
numpy
opencv-python
pillow
requests
ultralytics
pydicom
highdicom